+++ docs/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 -------------
 
+- 1.1.18 `August 16, 2021`
+    * `cirro prepare_data` accepts multiple input datasets to better support multimodal data
+
+- 1.1.17.post4 `August 13, 2021`
+    * Fixed error in `cirro launch` that prevented h5ad files from loading
+
 - 1.1.17.post3 `August 13, 2021`
     * Compute complete differential expression results in `cirro prepare_data` using `Scanpy`_ or `Pegasus`_/`Cumulus`_